requirements.txt

fastapi[all]~=0.104.1
torch
torchvision~=0.15.2
requests~=2.31.0
Pillow~=10.0.1
redis~=5.0.1
rq~=1.15.1
python-multipart
matplotlib
opencv-python~=4.8.1.78
numpy~=1.26.0
